Agregadas alertas en view getVaca

+++ app/src/main/java/com/example/tpfinalmoviles/data/model/Cow.kt
@@ -18,7 +18,7 @@ data class Cow(
     @SerializedName("ultimaFechaParto")
     val ultimaFechaParto: String?,
     @SerializedName("cc")
-    val cc: Int?
+    val cc: Double?
 )
 
 /*
